subscriber: use a multiprocessing context
Rather than set up forkserver in manage.py or run_with_daphne, we'll
set up forkserver in a context as part of the overseer, and use that
context to create the subscriber processes.

The overseer is created early enough that it won't inherit anything it
shouldn't, and then the forkserver usage is limited to
roost_backend.subscribers rather than potentially affecting the whole
process tree.

Also cleaned up a stray dict in the overseer that was never used.