views.py: rework MessageView
MessageView helper method is now a classmethod that prepares the
response payload itself. Only one query is made to retrieve messages,
and based on how many messages are retrieved, the value for `isDone`
is determined.

Longer term plans are to use this in consumers.py to provide a
websocket-based message retrival mechanism.